main/requirements.txt
asgiref==3.4.1; python_version >= "3.6"
beautifulsoup4==4.10.0; python_full_version > "3.0.0"
certifi==2021.10.8; python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.6.0"
charset-normalizer==2.0.8; python_full_version >= "3.6.0" and python_version >= "3"
django==3.2.8; python_version >= "3.6"
djangorestframework==3.12.4; python_version >= "3.5"
idna==3.3; python_version >= "3.5" and python_full_version < "3.0.0" or python_full_version >= "3.6.0" and python_version >= "3.5"
mysqlclient==2.1.0; python_version >= "3.5"
pytz==2021.3; python_version >= "3.6"
requests==2.26.0; (python_version >= "2.7" and python_full_version < "3.0.0") or (python_full_version >= "3.6.0")
soupsieve==2.3.1; python_version >= "3.6" and python_full_version > "3.0.0"
sqlparse==0.4.2; python_version >= "3.6"
urllib3==1.26.7; python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.6.0" and python_version < "4"
